harmonized some integral types

+++ Pandemic/City.h
@@ -93,8 +93,9 @@ public:
 
 	// Returns a report on the city
 	std::string string();
+	
 	// Sum up total cube numbers in the city
-	int totalCubes();
+	size_t totalCubes();
 
 	// Called when a player enters the city - calls the player role's onEnter(City&)
 	void onEnter(Player& player);
